Use ./pants bootstrap-cache-key for more reliable setup caching (#6)
The setup cache includes symlinks to things outside the cache, such as the (system) Python executable. Thus, the cache should only be reused on machines where those parts of the system exactly match up. This switches from trying to compute an appropriate cache key externally to letting the pants bootstrap script do it itself, via the new `PANTS_BOOTSTRAP_TOOLS=1 ./pants bootstrap-cache-key` (pantsbuild/setup#128).

This is a breaking change, in that it requires updating the checked in pants scripts to support the new tool.

Fixes #5