Add a pyenv action.
I was using https://github.com/gabrielfalcao/pyenv-action but caching
did not work out of the box and my attempts to get it working failed.
A read of that action code also seemed to confirm it does not actually
cache; it just re-uses the PYENV_ROOT if you've managed to cache it for
it.